Redid ability ID handling to better suit what it's going to be used for and added functionality to the server packet handler to let it cast server-side spells.

+++ packet/PacketHandlerClient.java
@@ -78,7 +78,7 @@ public class PacketHandlerClient implements IPacketHandler
             			break;
             		case PacketIds.applyStats:
             			String username = PacketHelper.readString(data);
-            			int race = data.readByte();
+            			byte race = data.readByte();
             			float energy = data.readFloat();
             			LoECraftPack.statHandler.updatePlayerData(username, Race.values()[race], energy);
             			break;
